+++ src/routes/NotFound.svelte
@@ -1,0 +1,16 @@
+<script>
+
+</script>
+
+<div class="relative">
+  <div
+    class="absolute inset-0 bg-gradient-to-b from-blue-900 from-80% to-white to-30%"
+  ></div>
+
+  <div class="relative px-4 py-8 text-center text-white">
+    <h1
+      class="relative pt-15 mb-4 z-10 text-4xl text-center leading-none tracking-tight md:text-4xl lg:text-5xl text-white"
+    > 
+    Not found</h1>
+  </div>
+</div>
